requirements.txt
# Core API and Database
requests>=2.31.0
python-dotenv>=1.0.0
psycopg>=3.1.12
praw>=7.7.1
pytrends>=4.9.0

# Data Analysis and Visualization
plotly>=5.18.0
pandas>=2.1.3
kaleido>=0.2.1

# Testing
pytest>=7.4.3
flake8>=6.1.0
pylint>=3.0.2

# Message queue
pika>=1.3.2
flask

# ML/AI Dependencies
torch
torchvision
torchaudio
huggingface-hub
transformers
accelerate
bitsandbytes